# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Compscore application",
        "title": "Compscore API",
        "contact": {
            "name": "1nv8rzim",
            "url": "https://github.com/compsore/compscore/issues"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/scoreboard/round/{round}": {
            "get": {
                "description": "Get scoreboard for a given round",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scoreboard"
                ],
                "summary": "Get scoreboard for a given round",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Round number",
                        "name": "round",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.Scoreboard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/scoreboard/status/{team}/{check}": {
            "get": {
                "description": "Get status scoreboard for a given team and check",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scoreboard"
                ],
                "summary": "Get status scoreboard for a given team and check",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Check name",
                        "name": "check",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Team number",
                        "name": "team",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.Status"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/scoreboard/status/{team}/{check}/{round}": {
            "get": {
                "description": "Get status scoreboard for a given team and check from a given round",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scoreboard"
                ],
                "summary": "Get status scoreboard for a given team and check from a given round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Check name",
                        "name": "check",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Team number",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Round number",
                        "name": "round",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.Status"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/scoreboard/team/{team}": {
            "get": {
                "description": "Get scoreboard of a given team for the current round",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scoreboard"
                ],
                "summary": "Get scoreboard of a given team for the current round",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team number",
                        "name": "team",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.TeamScoreboard"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/scoreboard/team/{team}/{round}": {
            "get": {
                "description": "Get scoreboard of a given team for a given round",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scoreboard"
                ],
                "summary": "Get scoreboard of a given team for a given round",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team number",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Round number",
                        "name": "round",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.TeamScoreboard"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
        "/status/{team}/{check}/{round}": {
            "get": {
                "description": "Get the status of a team for a given check for a given round",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Get the status of a team for a given check for a given round",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Check name",
                        "name": "check",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Round number",
                        "name": "round",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Error": {
            "description": "response of login request",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.Status": {
            "description": "status of a check",
            "type": "object",
            "properties": {
                "edges": {
                    "type": "object",
                    "properties": {
                        "check": {},
                        "round": {},
                        "team": {}
                    }
                },
                "error": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "structs.Check": {
            "description": "Check",
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "structs.Scoreboard": {
            "description": "Main Scoreboard",
            "type": "object",
            "properties": {
                "checks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/structs.Check"
                    }
                },
                "round": {
                    "type": "integer"
                },
                "scores": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "structs.Status": {
            "description": "Status",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "round": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "structs.TeamScoreboard": {
            "description": "Team Scoreboard",
            "type": "object",
            "properties": {
                "checks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/structs.Check"
                    }
                },
                "round": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ServiceAuth": {
            "description": "JWT for authentication",
            "type": "apiKey",
            "name": "auth",
            "in": "cookie"
        }
    }
}